+++ internal/scm/github/github.go
@@ -725,48 +725,6 @@ func (g *Github) getPullRequests(ctx context.Context, branchName string, repos [
 	return prs, nil
 }
 
-func (g *Github) CommitThroughAPI(ctx context.Context) error {
-	query := `
-		mutation ($input: CreateCommitOnBranchInput!) {
-			createCommitOnBranch(input: $input) {
-				commit {
-				url
-				}
-			}
-		}
-		`
-	vars := `
-	{
-	"input": {
-		"branch": {
-			"repositoryNameWithOwner": "chrisstatham/multi-gitter",
-			"branchName": "APPSEC-1108"
-		},
-		"message": {
-			"headline": "Hello from GraphQL!"
-		},
-		"fileChanges": {
-			"additions": [
-				{
-					"path": "myfile.txt",
-					"contents": "SGVsbG8gZnJvbSBKQVZBIGFuZCBHcmFwaFFM"      
-				}
-			]
-		},
-		"expectedHeadOid": "cb65177542c9239b9773fb4f6e117c46108c32d2" 
-		}
-	}`
-
-	result := map[string]graphqlRepo{}
-
-	err := g.makeGraphQLRequest(ctx, query, vars, &result)
-	if err != nil {
-		fmt.Printf("FAil")
-	}
-
-	return nil
-}
-
 func (g *Github) loggedInUser(ctx context.Context) (string, error) {
 	g.userMutex.Lock()
 	defer g.userMutex.Unlock()
